refactor: move setupFormErrorContext() to BaseController
