+++ core/lib/Thelia/Controller/Admin/BaseAdminController.php
@@ -25,7 +25,6 @@ use Thelia\Core\Security\Exception\AuthenticationException;
 use Thelia\Core\Security\Exception\AuthorizationException;
 use Thelia\Core\Template\ParserInterface;
 use Thelia\Core\Template\TemplateDefinition;
-use Thelia\Form\BaseForm;
 use Thelia\Log\Tlog;
 use Thelia\Model\AdminLog;
 use Thelia\Model\ConfigQuery;
@@ -137,32 +136,6 @@ class BaseAdminController extends BaseController
                 '%error' => $exception->getMessage(),
             ],
         );
-    }
-
-    protected function setupFormErrorContext(string $action, string $error_message, ?BaseForm $form = null, ?\Exception $exception = null): void
-    {
-        // Log the error message
-        Tlog::getInstance()->error(
-            $this->translator->trans(
-                'Error during %action process : %error. Exception was %exc',
-                [
-                    '%action' => $action,
-                    '%error' => $error_message,
-                    '%exc' => $exception instanceof \Exception ? $exception->getMessage() : 'no exception',
-                ],
-            ),
-        );
-
-        if ($form instanceof BaseForm) {
-            // Mark the form as errored
-            $form->setErrorMessage($error_message);
-
-            // Pass it to the parser context
-            $this->getParserContext()->addForm($form);
-        }
-
-        // Pass the error message to the parser.
-        $this->getParserContext()->setGeneralError($error_message);
     }
 
     /**
